+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Piral Changelog
 
+## 0.14.10 (tbd)
+
+- Improved support and documentation for wrappers
+- Improved default metadata using `piletConfig` from *package.json* in pilet debug (#462)
+- Fixed typing of extension slot defaults
+- Added new prop `meta` to registered pages
+- Added support for *debug-meta.json* when using `pilet debug` (#462)
+
 ## 0.14.9 (January 21, 2022)
 
 - Improved error output in `piral-cli-webpack5`
