+++ src/tooling/piral-cli/src/apps/debug-piral.ts
@@ -14,6 +14,7 @@ import {
   log,
   config,
   normalizePublicUrl,
+  logDone,
 } from '../common';
 
 export interface DebugPiralOptions {
@@ -151,11 +152,15 @@ export async function debugPiral(baseDir = process.cwd(), options: DebugPiralOpt
     bundlerName,
   );
 
+  bundler.ready().then(() => logDone(`Ready!`));
+
   bundler.on((args) => {
     hooks.afterBuild?.({ ...args, root, publicUrl, externals, entryFiles, name, bundler });
   });
 
+  const { piral: piralInjector, ...otherInjectors } = krasConfig.injectors;
   const injectorConfig = {
+    ...piralInjector,
     active: true,
     handle: ['/'],
     publicUrl,
@@ -163,7 +168,7 @@ export async function debugPiral(baseDir = process.cwd(), options: DebugPiralOpt
   };
 
   krasConfig.map['/'] = '';
-  krasConfig.injectors = reorderInjectors(injectorName, injectorConfig, krasConfig.injectors);
+  krasConfig.injectors = reorderInjectors(injectorName, injectorConfig, otherInjectors);
 
   log('generalVerbose_0004', `Using kras with configuration: ${JSON.stringify(krasConfig, undefined, 2)}`);
